Implement uplink confirmation
I have added a flag for uplink confirmation in lora.h and lora.c, and am setting this flag based on the callback events TWR_CMWX1ZZABZ_EVENT_MESSAGE_CONFIRMED and TWR_CMWX1ZZABZ_EVENT_MESSAGE_NOT_CONFIRMED. I've made sure that the last_uplink_confirmed flag is being updated based on whether the last message was confirmed or not - so i reset the flag when a new message is sent.
I have intentionally not reset msg.flags in application.c to ensure that the flags persist across messages. But this may be wrong as it looks like we are instantiating a new message_t each time? I'm not sure if after every send the device power-cycles - but the fact that msg is set as a static variable within the scope of build_status_message() whould mean that msg should retain its value across different invocations of the function.