Fixing for review comments
Reply to review:

The lora_send() confirmed parameter will always be true in this implementation.
The only call to lora_send() calls it with "true" as the arg.

If we're considerign the uplink confirmed flag to be "message sent" (without considering an ack, yet) then, maybe the right functionality would be to set the flag as false first after the message is built with the last value,    then set it to true when lora_send() is called?

I've tried to implement this by reseting the flag once a msg object is built to send, then once the send is complete, setting the flag so that the next msg will get this new flag if the send function didn't throw errors. If the send function did error out, then the flag would remain false?

I still do think that putting it in the modem callback switch tree is the better option since those callbacks are only triggered for modem activity - in this case, a broadcast activity. While my prev implementation was slightly off, resetting the flag in TWR_CMWX1ZZABZ_EVENT_SEND_MESSAGE_START and setting it in TWR_CMWX1ZZABZ_EVENT_SEND_MESSAGE_DONE makes sense to me - I made the changes as per the paragraph above but have addedd comments that match what i've said in this paragraph, both can be looked at.